Remove authcontent view
