--- src/view/AuthContent.tsx
@@ -1,34 +1,0 @@
-import { Auth, User, signInAnonymously } from 'firebase/auth'
-import { useSignOut } from 'react-firebase-hooks/auth'
-import { useState } from 'react'
-import ChakraWriting from '../lib/firewrite/chakra/Writing'
-import { useNavigate } from 'react-router-dom'
-
-export default function AuthContentView ({
-  auth, user
-}: {
-  auth: Auth
-  user?: User | null
-}): JSX.Element {
-  const navigate = useNavigate()
-  const [createAccountLoading, setCreateAccountLoading] = useState(false)
-  const [createAccountError, setCreateAccountError] = useState<Error>()
-  const [signOut, signOutLoading, signOutError] = useSignOut(auth)
-  async function createAccount (): Promise<void> {
-    try {
-      setCreateAccountLoading(true)
-      await signInAnonymously(auth)
-      setCreateAccountLoading(false)
-    } catch (error) {
-      setCreateAccountError(error as Error)
-    }
-  }
-  if (user == null) {
-    return <ChakraWriting label='New Account' write={createAccount} loading={createAccountLoading} error={createAccountError} />
-  }
-  async function signOutToHome (): Promise<void> {
-    await signOut()
-    navigate('/')
-  }
-  return <ChakraWriting label='Sign Out' write={signOutToHome} loading={signOutLoading} error={signOutError} />
-}
